Martin herrera (#24)
* archivos modificados para trabajar (?
