+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/controllers/PropertyController.java
@@ -8,6 +8,8 @@ import com.Egg.Inmobiliaria.models.Offer;
 import com.Egg.Inmobiliaria.models.Property;
 import com.Egg.Inmobiliaria.models.User;
 import com.Egg.Inmobiliaria.services.PropertyService;
+import com.Egg.Inmobiliaria.services.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
