+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/controllers/imagePropertyController.java
@@ -1,8 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package com.Egg.Inmobiliaria.controllers;
 
 
@@ -16,10 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-/**
- *
- * @author cosas
- */
+
 @Controller
 @RequestMapping("/image")
 public class imagePropertyController {
