--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entities/ImageProperty.java
@@ -1,4 +1,0 @@
-package com.Egg.Inmobiliaria.entities;
-
-public class ImageProperty {
-}
