--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entities/ImageUser.java
@@ -1,4 +1,0 @@
-package com.Egg.Inmobiliaria.entities;
-
-public class ImageUser {
-}
