--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entities/User.java
@@ -1,87 +1,0 @@
-package com.Egg.Inmobiliaria.entities;
-
-import com.Egg.Inmobiliaria.enumerations.Role;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-import org.hibernate.annotations.GenericGenerator;
-
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name = "uuid", strategy = "uuid2")
-    private Long id;
-
-    private String name;
-    private String email;
-    private String password;
-    private Long dni;
-    
-    @OneToOne
-    private ImageUser image;
-    
-    @Enumerated(EnumType.STRING)
-    private Role rol;
-
-    public Long getId() {
-        return id;
-    }
-
-    public ImageUser getImage() {
-        return image;
-    }
-
-    public void setImage(ImageUser image) {
-        this.image = image;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public Long getDni() {
-        return dni;
-    }
-
-    public void setDni(Long dni) {
-        this.dni = dni;
-    }
-
-    public Role getRol() {
-        return rol;
-    }
-
-    public void setRol(Role rol) {
-        this.rol = rol;
-    }
-
-}
