--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entitys/Offer.java
@@ -1,4 +1,0 @@
-package com.Egg.Inmobiliaria.entitys;
-
-public class Offer {
-}
