--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/entitys/User.java
@@ -1,4 +1,0 @@
-package com.Egg.Inmobiliaria.entitys;
-
-public class User {
-}
