--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumerations/OfferStatus.java
@@ -1,5 +1,0 @@
-package com.Egg.Inmobiliaria.enumerations;
-
-public enum OfferStatus {
-    CLIENT_OFFER, ENTITY_REJECTED, ENTITY_ACCEPTED, INACTIVE_OFFER,CLIENT_ACCEPTED;
-}
