--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumerations/PropertyStatus.java
@@ -1,6 +1,0 @@
-package com.Egg.Inmobiliaria.enumerations;
-
-public enum PropertyStatus {
-    FOR_SALE, FOR_RENT,
-    BOTH_STATUS, NONE;
-}
