--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumerations/PropertyType.java
@@ -1,8 +1,0 @@
-package com.Egg.Inmobiliaria.enumerations;
-
-public enum PropertyType {
-    CASA, DEPARTAMENTO, CABAÑA, DUPLEX,
-    TRIPLEX, PH, COCHERA, DEPOSITO,
-    PISO, TERRENO, OFICINA,
-    LOTE, COMERCIO, CAMPO;
-}
