--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/enumerations/Role.java
@@ -1,5 +1,0 @@
-package com.Egg.Inmobiliaria.enumerations;
-
-public enum Role {
-    ADMIN, ENTITY, CLIENT, BOTHROLE;
-}
