+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/exceptions/MiException.java
@@ -1,1 +1,16 @@
+package com.Egg.Inmobiliaria.exceptions;
+ 
 
+public class MiException extends Exception{
+
+    
+
+    public MiException(String msg) {
+
+        super(msg);
+
+    }
+
+    
+
+}
