+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/models/ImageProperty.java
@@ -1,4 +1,75 @@
 package com.Egg.Inmobiliaria.models;
 
+import javax.persistence.Basic;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Lob;
+import javax.persistence.ManyToOne;
+import org.hibernate.annotations.GenericGenerator;
+
+
+@Entity
 public class ImageProperty {
+    
+    @Id
+    @GeneratedValue(generator = "uuid")
+    @GenericGenerator(name = "uuid", strategy = "uuid2")
+    
+    private String id;
+    private String mime;
+    private String name;
+    @ManyToOne
+    private Property property;
+    @Lob @Basic(fetch = FetchType.LAZY)
+    private byte[] container;
+
+    public ImageProperty() {
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getMime() {
+        return mime;
+    }
+
+    public void setMime(String mime) {
+        this.mime = mime;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Property getProperty() {
+        return property;
+    }
+
+    public void setProperty(Property property) {
+        this.property = property;
+    }
+
+    public byte[] getContainer() {
+        return container;
+    }
+
+    public void setContainer(byte[] container) {
+        this.container = container;
+    }
+    
+    
+    
+    
+    
 }
