+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/models/Property.java
@@ -2,7 +2,6 @@ package com.Egg.Inmobiliaria.models;
 
 import com.Egg.Inmobiliaria.enums.PropertyStatus;
 import com.Egg.Inmobiliaria.enums.PropertyType;
-
 import javax.persistence.*;
 import java.time.LocalDate;
 import java.util.List;
