+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/repositories/IOfferRepository.java
@@ -1,0 +1,11 @@
+package com.Egg.Inmobiliaria.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.Egg.Inmobiliaria.models.Offer;
+
+@Repository
+public interface IOfferRepository extends JpaRepository<Offer, Long>{
+    
+}
