+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/repositories/UserRepository.java
@@ -1,13 +1,14 @@
 package com.Egg.Inmobiliaria.repositories;
 
-import com.Egg.Inmobiliaria.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import com.Egg.Inmobiliaria.models.User;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, String>{
+public interface UserRepository extends JpaRepository<User, Long>{
 
     @Query("SELECT u FROM User u WHERE u.email = :email")
     public User buscarPorEmail(@Param("email") String email);
