+++ Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/repositories/imagePropertyRepository.java
@@ -1,0 +1,14 @@
+package com.Egg.Inmobiliaria.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.Egg.Inmobiliaria.models.ImageProperty;
+
+
+@Repository
+public interface imagePropertyRepository extends JpaRepository<ImageProperty, String> {
+    
+   
+    
+}
