--- Inmobiliaria/src/main/java/com/Egg/Inmobiliaria/repositorys/imagePropertyRepository.java
@@ -1,21 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.Egg.Inmobiliaria.repositorys;
-
-import com.Egg.Inmobiliaria.entitys.ImageProperty;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- *
- * @author cosas
- */
-@Repository
-public interface imagePropertyRepository extends JpaRepository<ImageProperty, String> {
-    
-   
-    
-}
